requirements_streamlit.txt
streamlit>=1.28.0
cellpose>=4.0.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.20.0
matplotlib>=3.3.0
scikit-image>=0.18.0
scipy>=1.7.0
pandas>=1.3.0
